requirements.txt
dj-database-url==0.4.1
django==1.9.9
djangorestframework==3.5.2
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.2
